fix(#2087): widen Build{List,Tuple,Dict} count operand from u8 to u32 (#2091)
Summary
-------
list/tuple/set/dict literals with >= 256 elements were silently truncated
to (count mod 256) elements with no error. A 300-element literal produced
len 44; 256 -> 0; 1000 -> 232.

Root cause: the element/pair count operand of the BuildList / BuildTuple /
BuildDict bytecode instructions was a u8, and the compiler computed it as
`items.len() as u8`, wrapping modulo 256 at compile time. The VM then
consumed only `count mod 256` registers and dropped the rest. (Set literals
go through BuildList internally, so they were affected too.) The register
file is already u32-wide (`type Reg = u32`), so addressing was never the
limit -- only the count operand.

What Changed
------------
- bytecode.rs: BuildList / BuildTuple / BuildDict count operand u8 -> u32.
- compiler.rs: compile_collection, compile_set_literal (fast path),
  compile_dict_literal (fast path) now compute `n` as Reg (u32) instead of
  `as u8`; the register-overflow guards already used checked_add on Reg.
  Type-params and class-bases BuildTuple sites updated to pass u32.
- vm.rs: BuildList / BuildTuple / BuildDict handlers iterate `0..*n`
  directly (n is now Reg) instead of `Reg::from(*n)`.
- optimizer.rs: split BuildString out of the shared BuildList|BuildTuple
  match arms in three liveness helpers (it keeps its u8 operand), and
  dropped the now-redundant `as u32` casts on the widened operands.

Self-review caught
------------------
- BuildString (f-strings) is NOT affected: it already falls back to chained
  BinOp(Add) for > 255 parts, so it has no truncation bug. Left it u8 and
  only separated its optimizer match arms to avoid an or-pattern type
  conflict; the working f-string lowering is untouched.
- Verified register addressing was never the limit (Reg = u32); n=5000
  literals produce correct results, so this is purely the count operand.

Test plan
---------
- All four literal types at n = 255/256/257/300/1000/5000 now match
  python3.12 exactly for both length and contents (verified by hand).
- Nested large literal ([[...300...]]) and 256 string-keyed dict verified.
- New parity fixture: tests/cases/compiler/test_large_literal_count.py.
- cargo test --release --bin pyrust, --test parity_compare, and
  --workspace all green; debug build runs the fixture with no abort.

+++ crates/pyrust/src/bytecode.rs
@@ -199,9 +199,9 @@ pub enum Insn {
     /// return None
     ReturnNone,
     /// R[dst] = [R[base], R[base+1], ..., R[base+n-1]]
-    BuildList(Reg, Reg, u8),
+    BuildList(Reg, Reg, u32),
     /// R[dst] = (R[base], R[base+1], ..., R[base+n-1])
-    BuildTuple(Reg, Reg, u8),
+    BuildTuple(Reg, Reg, u32),
     /// R[dst] = R[base] ++ R[base+1] ++ ... ++ R[base+n-1]
     /// Concatenates `n` consecutive `str` registers into a single string in one
     /// pass over a preallocated buffer. Emitted only by f-string lowering
@@ -224,7 +224,7 @@ pub enum Insn {
     /// in `unpack_slice_key` (issue #931).
     BuildSlice(Reg, Reg),
     /// R[dst] = {R[base]: R[base+1], R[base+2]: R[base+3], ...}  (n key-value pairs)
-    BuildDict(Reg, Reg, u8),
+    BuildDict(Reg, Reg, u32),
     /// R[base..base+n] = iter_values(R[src])
     Unpack(Reg, Reg, u32),
     /// Extended unpack: R[src] is an iterable; store first `before` elements into
